--- a/draftjs_exporter/wrapper_state.py
+++ b/draftjs_exporter/wrapper_state.py
@@ -161,9 +161,7 @@ class WrapperState(object):
                     DOM.append_child(self.stack.head().elt, wrapper_parent)
                 else:
                     # Otherwise we can append at the end of the last child.
-                    wrapper_parent = (
-                        self.stack.head().last_child
-                    )
+                    wrapper_parent = self.stack.head().last_child
 
                 DOM.append_child(wrapper_parent, new_wrapper.elt)